feat(pdf): enhance trial dog PDF generation with new scoring and rendering functions
- Added new rendering functions for `AjoajanPisteytys`, `Ansiopisteet`, and `Tappiopisteet` to improve PDF generation.
- Updated the `renderTrialDogPdf` function to include these new scoring components.
- Refactored the PDF rendering logic to utilize shared drawing utilities for consistent text formatting.
- Removed the deprecated `koe-erat` module to streamline the codebase.

Refs: BEJ-89
