+++ apps/web/lib/server/trials/internal/tappiopisteet.ts
@@ -1,0 +1,10 @@
+import type { PDFFont, PDFPage } from "pdf-lib";
+
+// Placeholder section for AJOK tappiopisteet.
+// Add coordinates and drawing calls here when the data mapping is ready.
+export function drawTrialDogPdfTappiopisteet(input: {
+  page: PDFPage;
+  font: PDFFont;
+}): void {
+  void input;
+}
